ensuring that libs can still access httpd.log from logger module moving modified stream to httpd module to ensure stream catches all flows where httpd is used in lev-api
